signal修正
